Fix the shared library
Linking client applications to the shared library fails with
compilers that do not automatically manage the symbols, like
with MSVC 2019.

It looks like that using two definitions to manage which
'Q_DECL_<...>' to use (if 'Q_DECL_EXPORT' or 'Q_DECL_IMPORT')
is hiding the imports on the client application side. When
nothing is defined (like on client applications), the
'Q_DECL_IMPORT' seems to become missing, leading to a link
error.

It also seems unnecessary and more complicated to use
two definitions, while we can use just one to achieve
the desired result.

Fixes #58