Add compatibility with QT_NO_KEYWORDS macro (#104)
The Qt documentation recommends[1] not using Qt keywords in the
public API of Qt libraries, specially the 'signals' and 'slots'
Qt keywords.

Without this, client applications that are compiled with the
'QT_NO_KEYWORDS'[2] macro (or that are compiled in conjunction
with other libraries that also defines the same Qt keywords)
will fail to build.

The 'QT_NO_SIGNALS_SLOTS_KEYWORDS'[1] macro assures that
the library is not using 'signals' and 'slots' Qt keywords,
since it will fail to build when they are used. Useful for
code sanity and for avoiding future pull requests that may
try to use these Qt keywords.

[1] https://doc.qt.io/qt-6/signalsandslots.html#signals-and-slots-in-qt-based-libraries
[2] https://doc.qt.io/qt-6/qtglobal.html#QT_NO_KEYWORDS